--- assets/sea-config.json
@@ -1,5 +1,0 @@
-{
-    "main": "out.js",
-    "output": "sea-prep.blob",
-    "disableExperimentalSEAWarning": true
-}
